src_clue_ver1/requirements.txt
# Simplified Prototype Segmentation Requirements
# Install with: pip install -r requirements.txt

# Core PyTorch
torch>=1.9.0
torchvision>=0.10.0

# Data processing
numpy>=1.21.0
Pillow>=8.3.0

# Configuration
PyYAML>=5.4.0

# Progress bars
tqdm>=4.62.0

# Optional: For better performance
# CUDA support (uncomment if you have CUDA)
# torch-audio>=0.9.0

# Optional: For visualization (uncomment if needed)
# matplotlib>=3.4.0
# tensorboard>=2.7.0
